+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Fix an error for `RSpec/Extra/RestrictBlockTag`. ([@ydah])
+
 ## 0.1.0 - 2023-04-18
 
 - Initial release